% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/related_keywords.R
\name{related_keywords}
\alias{related_keywords}
\title{Related Keywords -
This function provides an extended list of related keywords, synonyms, and variations relevant to a queried term in a chosen database.}
\usage{
related_keywords(keyword, country, nb_of_keywords, semrush_Api_key)
}
\arguments{
\item{keyword}{A keyword or keyword expression you'd like to investigate.}

\item{country}{A regional database. Eg fr, es, de, us, uk, ru}

\item{nb_of_keywords}{The number of results returned to a request}

\item{semrush_Api_key}{An identification key assigned to a user after subscribing to SEMrush that is available via Profile page.}
}
\value{
related_keywords provides an extended list of related keywords, synonyms, and variations relevant to a queried
}
\description{
Related Keywords -
This function provides an extended list of related keywords, synonyms, and variations relevant to a queried term in a chosen database.
}
\examples{
\dontrun{
related_keywords("frank ocean","fr",10,"xxxxxxxxxxxxxxxxxxx")
}

}
